* BUGFIX: в виджетах выбора класса сортировка могла отличаться от естественной из-за того что JS сам сортирует ассоциативные массивы с целочисленными ключами. Теперь ключи жетско заданы как строки, впоследствии для вывода приводятся к целому числу

+++ scripts/schoolClassWidget.js
@@ -2,11 +2,12 @@ var updateClassList = function (classes, schoolSelectorEl, classSelectorEl)
 {
     classSelectorEl.set('html', '');
     var schoolID = schoolSelectorEl.value;
+    
     for(classID in classes[schoolID])
     {
         new Element('option',
         {
-            'value' : classID,
+            'value' : classID.toInt(),
             'text'  : classes[schoolID][classID]
         }).inject(classSelectorEl);
     }
